Add constants for remaining regs in datasheet - The datasheet leaves out a lot of information like which fields are which bit positions within a register, so this isn't set in stone
Implement the mfr's concept of banks as part of the register address (bit 8 is 0 or 1). Route things in readReg() and writeReg()

Add a port of existing code from another project on github (https://github.com/eouia/MMM-GroveGestures) just to test the code. Not working, ATM. Most likely what's missing is some key bit of initialization, also contained in that project (init(), along with the binary blob)